+++ tests/test_core.py
@@ -7,6 +7,7 @@ TODO:
 import os
 import slurmjobs
 import pathtree
+import pytest
 
 ROOT = os.path.dirname(__file__)
 
@@ -83,9 +84,10 @@ def test_shell():
 
 def test_arg_format():
     Args = slurmjobs.args.Argument
-    assert Args.get('fire') == slurmjobs.args.FireArgument
-    assert Args.get('argparse') == slurmjobs.args.ArgparseArgument
-    assert Args.get('sacred') == slurmjobs.args.SacredArgument
+    assert isinstance(Args.get('fire'), slurmjobs.args.FireArgument)
+    assert isinstance(Args.get('argparse'), slurmjobs.args.ArgparseArgument)
+    assert isinstance(Args.get('sacred'), slurmjobs.args.SacredArgument)
+    assert isinstance(Args.get('hydra'), slurmjobs.args.HydraArgument)
 
     CMD = 'python blah.py'
     CMD_ = CMD + ' {__all__}'
@@ -110,6 +112,43 @@ def test_arg_format():
         CMD_, arg1=['a', 'b'], arg2={'a': 5, 'b': 6, 'c': [1, 'b']})
     print('fire', cmd)
     assert cmd == CMD + ' --arg1=\'["a", "b"]\' --arg2=\'{"a": 5, "b": 6, "c": [1, "b"]}\''
+
+    cmd = Args.get('hydra').build(
+        CMD_, arg1=['a', 'b'], arg2={'a': 5, 'b': 6, 'c': [1, 'b']})
+    print('hydra', cmd)
+    assert cmd == CMD + ' +arg1=\'["a", "b"]\' +arg2=\'{"a": 5, "b": 6, "c": [1, "b"]}\''
+
+    cmd = Args.get('hydra').build(
+        CMD_, **{
+            'x': 5,
+            'arg1.x': ['a', 'b'],
+            'arg2.y': {'a': 5, 'b': 6, 'c': [1, 'b']},
+            '++y': 6,
+        })
+    print('hydra', cmd)
+    assert cmd == CMD + ' +x=5 +arg1.x=\'["a", "b"]\' +arg2.y=\'{"a": 5, "b": 6, "c": [1, "b"]}\' ++y=6'
+
+    cmd = Args.get('hydra', '++').build(
+        CMD_, **{
+            'x': 5,
+            'arg1.x': ['a', 'b'],
+            'arg2.y': {'a': 5, 'b': 6, 'c': [1, 'b']},
+            '++y': 6,
+        })
+    print('hydra', cmd)
+    assert cmd == CMD + ' ++x=5 ++arg1.x=\'["a", "b"]\' ++arg2.y=\'{"a": 5, "b": 6, "c": [1, "b"]}\' ++y=6'
+
+
+    cmd = Args.get('hydra', '').build(
+        CMD_, **{
+            'x': 5,
+            'arg1.x': ['a', 'b'],
+            'arg2.y': {'a': 5, 'b': 6, 'c': [1, 'b']},
+            '++y': 6,
+        })
+    print('hydra', cmd)
+    assert cmd == CMD + ' x=5 arg1.x=\'["a", "b"]\' arg2.y=\'{"a": 5, "b": 6, "c": [1, "b"]}\' ++y=6'
+
 
 
 def test_multicmd_arg_format():
@@ -178,3 +217,76 @@ def test_parameter_grid():
         {'something': 2, 'nodes': (7, 8, 9), 'a': 1, 'b': 3, 'some_flag': True},
         {'something': 2, 'nodes': (7, 8, 9), 'a': 2, 'b': 5, 'some_flag': True},
     ]
+
+
+def do_something(fname, content='hi'):
+    with open(fname, 'w') as f:
+        f.write(content)
+
+def do_something_wrong():
+    raise RuntimeError('jkasjkasfdjkajkhajkdh')
+
+def test_receipts(tmpdir):
+    out_file = os.path.join(tmpdir, 'asdf')
+    slurmjobs.Receipt.ROOT_DIR = tmpdir
+
+    r = slurmjobs.Receipt(do_something, out_file)
+    r.clear()
+    
+    # creates a file cuz it's the first time it runs
+    assert not r.exists
+    slurmjobs.use_receipt(do_something)(out_file)
+    assert os.path.isfile(out_file)
+    assert r.exists
+
+    assert not os.path.exists('./receipts')
+
+    os.remove(out_file)
+    assert not os.path.isfile(out_file)
+
+    # the receipt is written so it shouldn't do it again
+    assert r.exists
+    slurmjobs.use_receipt(do_something)(out_file)
+    assert not os.path.isfile(out_file)
+    assert r.exists
+
+    # now it should
+    assert r.exists
+    r.clear()
+    assert not r.exists
+    slurmjobs.use_receipt(do_something)(out_file)
+    assert os.path.isfile(out_file)
+    assert r.exists
+
+    os.remove(out_file)
+    assert not os.path.isfile(out_file)
+
+    # test dry run
+    r.clear()
+    slurmjobs.use_receipt(do_something)(out_file, test=True)
+    assert not os.path.isfile(out_file)
+    assert not r.exists
+
+
+    # try function taht fails
+    with pytest.raises(RuntimeError):
+        slurmjobs.use_receipt(do_something_wrong)()
+        assert not r.exists
+
+
+    # creates a file
+    slurmjobs.use_receipt(do_something)(out_file)
+    assert os.path.isfile(out_file)
+
+    os.remove(out_file)
+    assert not os.path.isfile(out_file)
+
+    # doesn't run again
+    slurmjobs.use_receipt(do_something)(out_file)
+    assert not os.path.isfile(out_file)
+
+    # overwrites
+    slurmjobs.use_receipt(do_something)(out_file, overwrite_=True)
+    assert os.path.isfile(out_file)
+
+    assert set(r.meta) == {'duration_secs', 'time'}